requirements.txt
psycopg2-binary==2.9.9
pymongo==4.7.1
pydantic==2.7.1
pydantic-settings==2.3.4
python-dateutil==2.9.0
psutil==5.9.8

# Testing
pytest==8.2.0
pytest-cov==5.0.0

# Linting & type checking
ruff==0.4.4
mypy==1.10.0
